test: fold writability cases into util tests
